+++ react/features/toolbox/hooks.web.ts
@@ -35,8 +35,8 @@ import {
     getParticipantsPaneOpen,
     isParticipantsPaneEnabled
 } from '../participants-pane/functions';
-import PipTriggerButton from '../pip/components/web/PipTriggerButton';
 import { useParticipantPaneButton } from '../participants-pane/hooks.web';
+import PipTriggerButton from '../pip/components/web/PipTriggerButton';
 import { usePollsButton } from '../polls/hooks.web';
 import { addReactionToBuffer } from '../reactions/actions.any';
 import { toggleReactionsMenuVisibility } from '../reactions/actions.web';
@@ -282,12 +282,12 @@ function useInviteButton() {
 }
 
 /**
- * Hide PiP toggle button when browser supports netiher Document PiP nor Video PiP (eg: firefox)
+ * Hide PiP toggle button when browser supports netiher Document PiP nor Video PiP (eg: firefox).
  *
  * @returns {Object | undefined}
  */
 function usePipToggleButton() {
-    //TODO: add support for Video PiP fallback. Hide only when both are not supported
+    // TODO: add support for Video PiP fallback. Hide only when both are not supported
     if ('documentPictureInPicture' in window) {
         return togglePiP;
     }
